Remove conversion to IamDataFrame
