+++ src/openscm_runner/adapters/base.py
@@ -36,3 +36,29 @@ class _Adapter(ABC):  # pylint: disable=too-few-public-methods
         **kwargs
             Keyword arguments used to initialise the model
         """
+
+    def run(self, scenarios, cfgs, output_variables):
+        """
+        Parameters
+        ----------
+        scenarios : :obj:`pyam.IamDataFrame`
+            Scenarios to run
+
+        cfgs : list[dict]
+            The config with which to run the model
+
+        output_variables : list[str]
+            Variables to include in the output
+
+        Returns
+        -------
+        :obj:`ScmDataFrame`
+            Model output
+        """
+        return self._run(scenarios, cfgs, output_variables)
+
+    @abstractmethod
+    def _run(self, scenarios, cfgs, output_variables):
+        """
+        Implementation of :meth:`run` interface
+        """
