+++ src/openscm_runner/adapters/magicc7/_run_magicc_parallel.py
@@ -142,6 +142,7 @@ def run_magicc_parallel(
 
     finally:
         instances.cleanup()
+        LOGGER.info("Shutting down parallel pool")
         shared_manager.shutdown()
         pool.shutdown()
